Add credit reader and current credits to main dialog.
At this point the user can now enter coins and see the total value. P.S.
Git's branch history is just hosed. I am not submitting half of these
files for the first time.

+++ src/vendingsimulation/displayui/MainDialogModel.java
@@ -20,7 +20,7 @@ public interface MainDialogModel
      * @param voltage_from_credit_reader The voltage received from the
      * credit reader, which defines the type of credit inserted
      */
-    void creditInserted( double voltage_from_credit_reader );
+    void creditInserted( BigDecimal voltage_from_credit_reader );
     
     /**
      * Handle when the customer requests inserted credits be returned.
